% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_phenotypes_generation.R
\name{combinatorial_phenotype_counts}
\alias{combinatorial_phenotype_counts}
\title{Count cells for each phenotype for each sample}
\usage{
combinatorial_phenotype_counts(
  processed_cell_data,
  parent_phen = NULL,
  min_count = 10,
  max_phenotype_length = 0,
  efficient = FALSE,
  n_threads = 1
)
}
\arguments{
\item{processed_cell_data}{Data.Frame containing filtered and thresholded cell data.

Use function \code{process_cell_data} to generate this input.}

\item{parent_phen}{Parent phenotype to filter for. All phenotypes generated will contain the parent phenotype.}

\item{min_count}{Minimum number of cells that a phenotype must have for at least one sample.}

\item{max_phenotype_length}{Maximum length of markers to compose a phenotype.}

\item{efficient}{If TRUE, filter full-length phenotypes for \code{min_count} condition before generating all marker combinations.
It is less sensitive for very rare phenotypes but yields a great boost in performance.}

\item{n_threads}{Number of threads to be used. Default: 1.}
}
\value{
Data.Frame with all possible phenotypes and cell counts for each sample.
}
\description{
\code{combinatorial_phenotype_counts} generates phenotypes considering all
possible combination of markers and count the number of cells for each phenotype
for each sample.
}
