% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_comparisons.R
\name{compute_statistically_relevant_phenotypes}
\alias{compute_statistically_relevant_phenotypes}
\title{Filter statistically relevant phenotypes by comparing two sample groups}
\usage{
compute_statistically_relevant_phenotypes(
  phenotype_cell_counts,
  channel_data,
  sample_data,
  groups_column,
  g1,
  g2,
  max_pval = 0.05,
  parent_phen = NULL,
  n_threads = 1
)
}
\arguments{
\item{phenotype_cell_counts}{Data.Frame object with marker and sample columns. Each row
is a unique phenotype.}

\item{channel_data}{Data.Frame containing columns named: Channel, Marker, T1, [T2, T3, ... , Tn], [OOB].}

\item{sample_data}{Data.Frame containing a Sample_ID column and additional grouping columns for the samples.}

\item{groups_column}{Column name in \code{sample_data} where group identifications are stored.}

\item{g1}{Group label or vector with group labels for first group.}

\item{g2}{Group label or vector with group labels for second group.}

\item{max_pval}{Maximum p-value. Used to filter phenotypes in final output.}

\item{parent_phen}{Parent phenotype to filter for. All phenotypes in the output will contain the parent phenotype.}

\item{n_threads}{Number of threads to be used. Default: 1}
}
\description{
\code{compute_statistically_relevant_phenotypes} filters statistically
relevant phenotypes by comparing two sample groups defined in \code{sample_data}.
}
