% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_markers}
\alias{count_markers}
\title{Count markers in phenotype}
\usage{
count_markers(phenotype)
}
\arguments{
\item{phenotype}{Numeric vector with phenotype encoding.}
}
\value{
Number of markers.
}
\description{
\code{count_markers} counts the number of markers a phenotype has
given its numerical representation. Basically it returns the number of
values greater than -1 in an integer vector.
}
\examples{
count_markers(c(0,1,2,0))
count_markers(c(0,1,-1))

}
