% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csv_read_n_sorted_phenotypes}
\alias{csv_read_n_sorted_phenotypes}
\title{Get n_phenotypes from file sorting by sorted_column}
\usage{
csv_read_n_sorted_phenotypes(
  file_path,
  n_phenotypes,
  sorted_column,
  decreasing = FALSE
)
}
\description{
Get n_phenotypes from file sorting by sorted_column
}
