% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_phenotype_in_file}
\alias{find_phenotype_in_file}
\title{Find phenotype in csv file}
\usage{
find_phenotype_in_file(
  phen_file,
  phen,
  markers,
  n_cores = 1,
  chunk_size = 10000
)
}
\description{
Find phenotype in csv file
}
