% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_phenotype_identification.R
\name{get_independent_relevant_phenotypes}
\alias{get_independent_relevant_phenotypes}
\title{Identify independent relevant phenotypes}
\usage{
get_independent_relevant_phenotypes(
  phen_data,
  channel_data,
  n_phenotypes = 1000,
  min_confidence = 0.5,
  n_threads = 1
)
}
\arguments{
\item{phen_data}{A Data.Frame with Marker columns, sample columns, and (effect_size, p_value, log2foldChange) columns.}

\item{channel_data}{Data.Frame containing columns named: Channel, Marker, T1, [T2, T3, ... , Tn], [OOB].}

\item{n_phenotypes}{maximum number of phenotypes to be considered from \code{phen_data} filtered by lowest p-values. Default: 1000.}

\item{min_confidence}{Minimal confidence threshold to filter output. Default: 0.5.}

\item{n_threads}{Number of threads to be used. Default: 1.}
}
\description{
It uses a network clustering process to identify most representative
phenotypes when several similar phenotypes are relevant.
}
