% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_phenotype_name}
\alias{make_phenotype_name}
\title{Generates a human-readable phenotype name}
\usage{
make_phenotype_name(phenotype, markers)
}
\arguments{
\item{phenotype}{Numeric vector with phenotype encoding.}

\item{markers}{Names for markers. Must match the length of \code{phenotype}.}
}
\value{
Phenotype name.
}
\description{
\code{make_phenotype_name} generates a human-readable phenotype name from a numeric representation
with the following encoding:
-1 : Drop marker
0  : - (negative)
1  : + (positive)
2  : ++ (double positive)
More positive states can also be represented (e.g. 3, 4... : +++, ++++...)
}
\examples{
make_phenotype_name(c(0,1,2,0),c("Marker1","Marker2","Marker3","Marker4"))
make_phenotype_name(c(0,1,-1),c("Marker1","Marker2","Marker3"))

}
