% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phenotype_to_numbers}
\alias{phenotype_to_numbers}
\title{Encode phenotype name}
\usage{
phenotype_to_numbers(phenotype, markers)
}
\arguments{
\item{phenotype}{Phenotype name.}

\item{markers}{Names of the markers.}
}
\value{
Numeric vector with encoded phenotype.
}
\description{
\code{phenotype_to_numbers} generates a numeric representation of a human-readable phenotype name
using the following encoding:
-1 : Droped marker
0  : - (negative)
1  : + (positive)
2  : ++ (double positive)
More positive states can also be represented (e.g. 3, 4... : +++, ++++...)
}
\examples{
phenotype_to_numbers("Marker1-Marker2+Marker3++Marker4-",c("Marker1","Marker2","Marker3","Marker4"))
phenotype_to_numbers("Marker1-Marker2+",c("Marker1","Marker2","Marker3"))

}
