% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_phenotypes_generation.R
\name{process_cell_data}
\alias{process_cell_data}
\title{Process raw cell data}
\usage{
process_cell_data(
  cell_data,
  channel_data,
  sample_data,
  sampleID_col = "Sample_ID",
  n_threads = 1
)
}
\arguments{
\item{cell_data}{Data.Frame where each column is a channel of measurement plus one column with sample IDs. Each row is a cell.}

\item{channel_data}{Data.Frame containing columns named: Channel, Marker, T1, [T2, T3, ... , Tn], [OOB].}

\item{sample_data}{Data.Frame containing a Sample_ID column and additional grouping columns for the samples.}

\item{n_threads}{Number of threads to be used (Only useful if there is an OOB column in channel_data). Default: 1}

\item{sample_ID_col}{Name of the column in \code{cell_data} where the Sample IDs are stored. Default: "Sample_ID"}
}
\value{
Data.Frame with filtered and discretized cells.
}
\description{
\code{process_cell_data} uses the channel_data and sample_data to filter
columns and cells from cell_data and discretize values based on thresholds
defined on channel_data.
}
