% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_comparisons_server.R
\name{statistically_relevant_phenotypes_server}
\alias{statistically_relevant_phenotypes_server}
\title{Filter statistically relevant phenotypes by comparing two sample groups (Server Version)}
\usage{
statistically_relevant_phenotypes_server(
  output_folder,
  channel_file,
  sample_file,
  input_phenotype_counts = "combinatorial_phenotype_counts.csv",
  output_file = "significant_phenotypes.csv",
  log_file = "significant_phenotypes.log",
  test_type = "group",
  groups_column = NULL,
  g1 = NULL,
  g2 = NULL,
  correlation_column = NULL,
  survival_time_column = NULL,
  survival_status = NULL,
  max_pval = 0.05,
  parent_phen = NULL,
  continue = FALSE,
  n_threads = 1,
  verbose = FALSE
)
}
\arguments{
\item{output_folder}{Path to folder where output files from this and previous steps should be saved.}

\item{channel_file}{Path to a ".csv" file containing columns named: Channel, Marker, T1, [T2, T3, ... , Tn], [OOB].}

\item{sample_file}{Path to a ".csv" file containing a Sample_ID column and additional grouping columns for the samples.}

\item{test_type}{Type of statistical test to be performed. Value can be "group", "correlation", or "survival". Default: "group".
Additional parameters should be provided accordingly, such as [groups_column, g1, g2] for "group", [correlation_column] for "correlation", 
and [survival_time_column, survival_status_column] for "survival". Parameters not used in the test will be ignored.}

\item{groups_column}{Column name in \code{sample_data} where group identifications are stored.}

\item{g1}{Group label or vector with group labels for first group.}

\item{g2}{Group label or vector with group labels for second group.}

\item{correlation_column}{Column name in \code{sample_data} where data to be correlated is stored.}

\item{survival_time_column}{Column name in \code{sample_data} where survival time is stored.}

\item{max_pval}{Maximum p-value. Used to filter phenotypes in final output.}

\item{parent_phen}{Parent phenotype to filter for. All phenotypes in the output will contain the parent phenotype.}

\item{continue}{If TRUE, look for files to resume execution. Also needed to save necessary "continuing" files.}

\item{n_threads}{Number of threads to be used. Default: 1.}

\item{verbose}{If TRUE, print outputs from log to stdout.}

\item{survival_status_column}{Column name in \code{sample_data} where survival status is stored.}
}
\value{
Output is saved to file.
}
\description{
\code{compute_statistically_relevant_phenotypes} filters statistically
relevant phenotypes by comparing two sample groups defined in \code{sample_data}.
}
\details{
This version reads the "output_folder/combinatorial_phenotype_counts.csv" by chunks
and save the partial results to a file. It is intended to be used with large datasets
which inputs wouldn't fit in the available memory.
}
